+++ packages/gamut-labs/src/GlobalHeader/types.tsx
@@ -10,7 +10,10 @@ type RenderProfile = { desktop: () => ReactNode; mobile: () => ReactNode };
 type RenderSearch = { desktop: () => ReactNode; mobile: () => ReactNode };
 
 type BaseHeader = {
+  /** A method to be called on click/activating a header item */
   action: AppHeaderClickHandler;
+  /** A method to be called only on click/activating a *link* header item */
+  onLinkAction?: AppHeaderClickHandler;
   className?: string;
   renderSearch?: RenderSearch;
 };